the third player for those that lose count, this is a scrubbed version (with a librified back-end and some other thin...
Original commit message from CVS:

the third player
for those that lose count, this is a scrubbed version (with a librified
back-end and some other things) of Company's hack of arik's rework of
the original player

also, a spec file is included which works (but too much gnome2 is drawn in,
like ssl) and the pixmaps should be visible whether you run it uninstalled
or installed.

hope that makes you all feel warm inside.