gtk-doc-plugins.mak: Add 'check-inspected-versions' target; this helps identify files that should have been removed o...
Original commit message from CVS:
* gtk-doc-plugins.mak:
Add 'check-inspected-versions' target; this helps identify
files that should have been removed or where the version
number should (ideally) be updated before a release
(which doesn't happen automatically if the releaser doesn't
build that plugin locally). Not adding at a distcheck hook
yet though, because it's not really that important and would
probably also be a problem on buildbots.