build: use glib_gen_basename for generated files. Fixes #598114
We used pattern rules before. This breaks if a header is references from a
different directory. For marshallers I could trigger useless rebuilds. For enums
it caused that the genrated files where overwritten with wrong content as the
pattern rule has a static dependecy. So it created enumtypes for dir a in dir b.