Add cruft.mak file that can be included from modules to check for cruft
When plugins get renamed, moved or removed we often leave behind unwanted
left-overs in our source tree, since no one runs 'make clean' before every
git update. This can cause problems and be confusing, so add a little
Makefile rule to include to print a warning for if any of the files or
directories specified in CRUFT_FILES and/or CRUFT_DIRS are still present
in our uninstalled tree.