Split ERROR_{C|CXX} flags into WARNING and ERROR flags
The WARNING flags now contain all warning flags, the ERROR flags contain
the flags related to making warnings fatal. For more explanation, see
the inline docs.

The rationale is that we want to turn on a lot more warning flags, but
generated and imported sources might not have the same quality. For
those, we want to use the WARNING_CFLAGS but not the ERROR_CFLAGS.