upload-doc.mak: upload release docs to the right place
On our website, all doc links point to the docs for head, which
is where git docs get uploaded to, while release docs for some
reason get uploaded to docs/$version where no one will ever
find them. Change upload rule to upload release docs to both
the versioned directory and head, so that new docs will actually
be visible on the website when the new-release script in www
uploads them as part of the release process.