gstdoc-scangobj: serialise doubles and floats always with a decimal dot
Make sure floats and double property values are output with a decimal dot
(and not e.g. a comma) irrespective of the current locale. g_ascii_formatd() is
used here instead of g_ascii_dtostr() because we want nice human-readable
numbers and do not need the exact same bit representation when deserialising.

Other parts of gtk-doc may need fixing as well to make sure to always
deserialise floats and doubles in C locale.