Revert "gtk-doc.mak: make build break on new api that has not been added to the sections file"
This reverts commit 30c7a70b1d4765d26789950ca8e983284c0a05b3.

Docs stuff should never break the build (it's fragile enough as it is), especially
not when whether things break or not depends on factors outside of the source tree
such as, for example, the gtk-doc version being used. Also, hooks like this are
probably more appropriate for the check or distcheck target.

https://bugzilla.gnome.org/show_bug.cgi?id=625862