m4: add AG_GST_PKG_CONFIG_PATH macro
Sets up a GST_PKG_CONFIG_PATH variable for use in Makefile.am
which contains the path of the in-tree pkgconfig directory first
and then any paths specified in PKG_CONFIG_PATH.

We do this mostly so we don't have to use unportable shell constructs
such as ${PKG_CONFIG_PATH:+:$PKG_CONFIG_PATH} in Makefile.am to handle
the case where the environment variable is not set, but also in order
to avoid a trailing ':' in the PKG_CONFIG_PATH which apparently causes
problems with pkg-config on windows with msys/mingw.

Also shows the PKG_CONFIG_PATH set up in the configure output,
which is handy when debugging other people's configure problems.