gst-autogen.sh: drop support for AUTORECONF env var
autoreconf is hard-coded now after recent changes, so there's
no point in checking for AUTORECONF if we're not going to
use it anyway. It's not a standard environment variable
anyway, nor should it be required for anything (autoreconf
itself supports the environment variables AUTOM4TE, AUTOCONF,
AUTOHEADER, AUTOMAKE, ACLOCAL, AUTOPOINT, LIBTOOLIZE, M4,
and MAKE).

https://bugzilla.gnome.org/show_bug.cgi?id=709480