check.mak: parallelise 'check-valgrind' target
If we have multiple cores or CPUs, let's use all of them for
running the valgrind checks in parallel!

The output is not so nice any more now, but if there's a
problem we'll see from the stack trace in the valgrind output
where it comes from anyway, so not really a problem.

And people can still use make -j1 of course.