autogen.sh.in: simplify checks for autoconf and pkg-config
No need to check the versions there. pkg-config 0.8 was
released in 2003. autoconf version checks will be made
by autoreconf based on our requirement in configure.ac,
and it will error out for us in that case.

So just check for presence of these tools and get rid
of a whole lot of code.