Pre-commit: Fallback to copy when symlink does not work
On certain filesystems like FAT, NTFS and some network file systems,
symlinks are not supported. In those case, installing the pre-commit
hook fails and leaves the developers with having to run the style
checker manually.

Instead, fallback to copy. This should enable the commit hook for
users of filesystems and certain OSs where symlinks are not supported.

https://bugzilla.gnome.org/show_bug.cgi?id=759089