use brace expansion, not command substitution
Recently noticed that gtk-doc stopped to extract the internal
documentation (appeared just API, but no description).

The problem seems to be a regression from commit f363b32, where the
expasion of the variable _source_dir is expanded using parentheses,
which are used for command substitution, when the variable need to
be a string expansion, and for that brace are needed by the shell.

https://bugzilla.gnome.org/show_bug.cgi?id=768571